fix: resolve critical numerical bugs and modernize for Python 3 / NumPy 2.0
This release fixes several critical bugs that caused incorrect simulation
results and updates the codebase for compatibility with modern Python
ecosystem (Python 3.8+, NumPy 2.0, h5py 3.0, SciPy 1.17).

Critical numerical fixes:
- Fix finite difference operators in flux calculations (column.py)
  The 2nd-order formulas incorrectly used multiplication instead of
  subtraction, producing wrong boundary flux values
- Fix array indexing bug: C[-3, :] → C[-3, idx] in bottom flux

Python 3 compatibility:
- Fix exec() local scope handling in create_dynamic_functions()
- Replace raw_input() with input() in vg.py
- Replace string identity checks (is/is not) with equality (==/!=)
- Replace bare except with specific exception types

Library compatibility:
- NumPy 2.0: Replace deprecated np.cumproduct with np.cumprod
- NumPy 1.18+: Cast float to int for np.linspace num parameter
- h5py 3.0: Use public API (h5py.Dataset) and item[()] accessor
- SciPy: Fix ODE function return shape (1D array required)

Other fixes:
- Fix NotImplemented → NotImplementedError() exception type

Bump version to 1.4.2

+++ tests/general_tests.py
@@ -50,7 +50,7 @@ class TestMathModel:
         lab.add_species(True, 'O2', D, 0, bc_top=1, bc_top_type='dirichlet', bc_bot=0, bc_bot_type='neumann')
         lab.dcdt.O2 = '0'
         lab.solve()
-        x = np.linspace(0, lab.length, lab.length / lab.dx + 1)
+        x = np.linspace(0, lab.length, int(lab.length / lab.dx) + 1)
         sol = 1 / 2 * (special.erfc((x - lab.w * lab.tend) / 2 / np.sqrt(D * lab.tend)) + np.exp(
             lab.w * x / D) * special.erfc((x + lab.w * lab.tend) / 2 / np.sqrt(D * lab.tend)))
